Move db module to cmd/root module
If I try to keep the db module in `db/db.go`, then I run into
circular import statements when I was to use Reps, and db calls.
