+++ cmd/root/db_test.go
@@ -1,4 +1,4 @@
-package db
+package root
 
 import (
 	"os"
